requirements/base.txt
bs4
ccxt
cryptocompare==0.6
cryptography==2.1.4
django==2.0.5
django-bootstrap3==8.2.3
django-filter
django-ratelimit
djangorestframework==3.7.3
gitterpy
gunicorn
matplotlib
markdown
reportlab
requests
requests_oauthlib
Pillow==4.0.0
premailer
populus
psycopg2
PyJWT==1.5.3
PyPdf2
python-twitter==3.2
sendgrid
slackclient
web3==4.0.0b10
mailchimp3
django_chartit==0.2.9
MechanicalSoup==0.9.0.post4
pyshorteners
statistics
django-email-obfuscator
cytoolz==0.9.0
boto==2.48.0
rollbar==0.13.18
awscli
google-api-python-client
whitenoise==3.3.1
django-environ==0.4.4
ipfsapi
eth-utils==1.0.0b2
jsondiff==1.1.1
social-auth-app-django==2.1.0
django-ipware==2.0.2
geoip2==2.8.0
django-silk==2.0.0
django-extensions==2.0.6
ipdb==0.11
inotify==0.2.9
matplotlib
imageio
boto3
django-storages
easy-thumbnails
django-taggit==0.22.2
django-crispy-forms==1.7.2
